+++ index.html
@@ -212,7 +212,7 @@
 </div>
 
 
-<script src="lib/js/index.js"></script>
+<script src="main.bundle.js"></script>
 
 </body>
 </html>
