+++ lib/index.js
@@ -1,4 +1,6 @@
 // food starts here
+var Breakfast = require('./breakfast.js')
+
 var allFoods = document.getElementById('myTable')
 
 function submitFood(food, calories, id){
@@ -74,38 +76,32 @@ function updateExercise() {
 
 updateExercise();
 
-//
-//
-// start of breakfast
-//
-//
 
-// stores checked items one by one into localStorage
-function storeBreakfast(food, calories){
-  var breakfastJSON = localStorage.getItem('breakfast')
+// START OF BREAKFAST
+$('#submit-breakfast').on('click', function(event){
+  event.preventDefault();
+  var checkBox = $('.food-check')
+  breakfastChecker(checkBox);
+  $('#totals > tr').remove()
+  updateTotalTable()
+})
 
-  if(breakfastJSON === null){
-    breakfastJSON = '[]';
-  };
+$('#breakfastTable').on("click", ".fa-trash-o", function(event){
+  that = this;
+  var breakfastClass = new Breakfast()
+  breakfastClass.deleteBreakfast(event, that)
+  $('#totals > tr').remove()
+  updateTotalTable()
+});
 
-  var currentBreakfast = JSON.parse(breakfastJSON);
-  currentBreakfast.unshift({food: food, calories: calories});
-  breakfastJSON = JSON.stringify(currentBreakfast);
-  localStorage.setItem('breakfast', breakfastJSON)
 
-  $("#breakfastTable > tr").remove();
-  updateBreakfast();
-
-}
-
-// grabs items if the checkbox is checked and sends to store breakfast
 function breakfastChecker(checkBox){
   for (var i = 0; i < checkBox.length; i++) {
     if (checkBox[i].checked === true){
       var food = checkBox[i].parentElement.parentElement.parentElement.children[0].innerText;
       var calories = checkBox[i].parentElement.parentElement.parentElement.children[1].innerText;
-      storeBreakfast(food, calories)
-
+      var breakfastClass = new Breakfast(food, calories)
+      breakfastClass.storeBreakfast(food, calories)
     }
   }
   clearChecked(checkBox);
@@ -162,9 +158,7 @@ function submitBreakfastCalories(total=0) {
 
   var breakfastTable = document.getElementById('breakfastTable')
   breakfastTable.appendChild(newRow);
-
   // remaining row
-
   var remainingRow = document.createElement('tr');
   remainingRow.id = "breakfast-calorie-total"
 
@@ -201,22 +195,7 @@ function updateBreakfast(){
   }
 }
 
-function deleteBreakfast(event, that) {
-  event.preventDefault();
-  var breakfast = JSON.parse(localStorage["breakfast"]);
-  for (var i = 0; i < breakfast.length; i++) {
-    if ($(that).parent().parent().data('id') === i){
-      that.parentElement.parentElement.remove();
-      breakfast.splice(i,1)
-      breakfastJSON = JSON.stringify(breakfast);
-      localStorage.setItem('breakfast', breakfastJSON);
-    }
-   }
-   $("#breakfastTable > tr").remove();
-   updateBreakfast();
- };
 
-updateBreakfast()
 
 // lunch starts here
 
@@ -825,22 +804,7 @@ function makeDate(){
 $(document).ready(function(){
 
   // breakfast doc ready starts here
-
-  $('#submit-breakfast').on('click', function(event){
-    event.preventDefault();
-    var checkBox = $('.food-check')
-    breakfastChecker(checkBox);
-    $('#totals > tr').remove()
-    updateTotalTable()
-  })
-
-  $('#breakfastTable').on("click", ".fa-trash-o", function(event){
-    that = this;
-    deleteBreakfast(event, that);
-    $('#totals > tr').remove()
-    updateTotalTable()
-  });
-
+  updateBreakfast()
   // lunch doc ready starts here
